whatsapp-service/requirements.txt
fastapi
uvicorn
pywhatkit
requests
python-dotenv
